refactor: 사용되지 않는 파일 삭제 및 용도별 그룹화
