+++ src/components/common/button/BattleButton.tsx
