+++ src/components/post/detail/index.tsx
@@ -3,10 +3,10 @@ import { useQueries } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 
+import Battle from '@/components/common/button/BattleButton';
+import Like from '@/components/common/button/LikeButton';
+import MusicPlayButton from '@/components/common/button/MusicPlayButton';
 import Header from '@/components/common/Header';
-import Battle from '@/components/common/ImageButtons/BattleButton';
-import Like from '@/components/common/ImageButtons/LikeButton';
-import MusicPlayButton from '@/components/common/MusicPlayButton';
 import useAuth from '@/components/login/useAuth';
 import { COLOR } from '@/constants/color';
